requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
line-bot-sdk>=3.11.0
anthropic>=0.28.0
httpx>=0.27.0
python-dotenv>=1.0.0
